+++ crates/stages/stages/benches/setup/mod.rs
@@ -6,6 +6,7 @@ use reth_db_api::{
     transaction::{DbTx, DbTxMut},
 };
 use reth_primitives::{Account, Address, SealedBlock, B256, U256};
+use reth_provider::StorageWriter;
 use reth_stages::{
     stages::{AccountHashingStage, StorageHashingStage},
     test_utils::{StorageKind, TestStageDB},
@@ -17,7 +18,7 @@ use reth_testing_utils::{
         random_eoa_accounts,
     },
 };
-use reth_trie::{updates::StorageWriter, StateRoot};
+use reth_trie::StateRoot;
 use std::{collections::BTreeMap, fs, path::Path, sync::Arc};
 use tokio::runtime::Handle;
 
@@ -139,7 +140,7 @@ pub(crate) fn txs_testdata(num_blocks: u64) -> TestStageDB {
         let offset = transitions.len() as u64;
 
         db.insert_changesets(transitions, None).unwrap();
-        db.commit(|tx| Ok(updates.flush(&StorageWriter, tx)?)).unwrap();
+        db.commit(|tx| Ok(StorageWriter.write_trie_updates(updates, tx)?)).unwrap();
 
         let (transitions, final_state) = random_changeset_range(
             &mut rng,
