+++ crates/stages/stages/src/stages/merkle.rs
@@ -8,15 +8,13 @@ use reth_db_api::{
 use reth_primitives::{BlockNumber, GotExpected, SealedHeader, B256};
 use reth_provider::{
     DatabaseProviderRW, HeaderProvider, ProviderError, StageCheckpointReader,
-    StageCheckpointWriter, StatsReader,
+    StageCheckpointWriter, StatsReader, StorageWriter,
 };
 use reth_stages_api::{
     BlockErrorKind, EntitiesCheckpoint, ExecInput, ExecOutput, MerkleCheckpoint, Stage,
     StageCheckpoint, StageError, StageId, UnwindInput, UnwindOutput,
 };
-use reth_trie::{
-    updates::StorageWriter, IntermediateStateRootState, StateRoot, StateRootProgress, StoredSubNode,
-};
+use reth_trie::{IntermediateStateRootState, StateRoot, StateRootProgress, StoredSubNode};
 use std::fmt::Debug;
 use tracing::*;
 
@@ -219,7 +217,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
                 })?;
             match progress {
                 StateRootProgress::Progress(state, hashed_entries_walked, updates) => {
-                    updates.flush(&StorageWriter, tx)?;
+                    StorageWriter.write_trie_updates(updates, tx)?;
 
                     let checkpoint = MerkleCheckpoint::new(
                         to_block,
@@ -239,7 +237,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
                     })
                 }
                 StateRootProgress::Complete(root, hashed_entries_walked, updates) => {
-                    updates.flush(&StorageWriter, tx)?;
+                    StorageWriter.write_trie_updates(updates, tx)?;
 
                     entities_checkpoint.processed += hashed_entries_walked as u64;
 
@@ -254,7 +252,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
                         error!(target: "sync::stages::merkle", %e, ?current_block_number, ?to_block, "Incremental state root failed! {INVALID_STATE_ROOT_ERROR_MESSAGE}");
                         StageError::Fatal(Box::new(e))
                     })?;
-            updates.flush(&StorageWriter, provider.tx_ref())?;
+            StorageWriter.write_trie_updates(updates, provider.tx_ref())?;
 
             let total_hashed_entries = (provider.count_entries::<tables::HashedAccounts>()? +
                 provider.count_entries::<tables::HashedStorages>()?)
@@ -327,7 +325,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
             validate_state_root(block_root, target.seal_slow(), input.unwind_to)?;
 
             // Validation passed, apply unwind changes to the database.
-            updates.flush(&StorageWriter, provider.tx_ref())?;
+            StorageWriter.write_trie_updates(updates, provider.tx_ref())?;
 
             // TODO(alexey): update entities checkpoint
         } else {
