+++ crates/storage/provider/src/bundle_state/bundle_state_with_receipts.rs
@@ -68,7 +68,7 @@ impl StateWriter for ExecutionOutcome {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{test_utils::create_test_provider_factory, AccountReader};
+    use crate::{test_utils::create_test_provider_factory, AccountReader, StorageWriter};
     use reth_db::test_utils::create_test_rw_db;
     use reth_db_api::{
         cursor::DbDupCursorRO,
@@ -78,7 +78,7 @@ mod tests {
     use reth_primitives::{
         keccak256, Account, Address, Receipt, Receipts, StorageEntry, B256, U256,
     };
-    use reth_trie::{test_utils::state_root, updates::StorageWriter, StateRoot};
+    use reth_trie::{test_utils::state_root, StateRoot};
     use revm::{
         db::{
             states::{
@@ -892,7 +892,7 @@ mod tests {
             }
 
             let (_, updates) = StateRoot::from_tx(tx).root_with_updates().unwrap();
-            updates.flush(&StorageWriter, tx).unwrap();
+            StorageWriter.write_trie_updates(updates, tx).unwrap();
         })
         .unwrap();
