+++ crates/trie/parallel/benches/root.rs
@@ -6,12 +6,11 @@ use rayon::ThreadPoolBuilder;
 use reth_primitives::{Account, B256, U256};
 use reth_provider::{
     bundle_state::HashedStateChanges, providers::ConsistentDbView,
-    test_utils::create_test_provider_factory,
+    test_utils::create_test_provider_factory, StorageWriter,
 };
 use reth_tasks::pool::BlockingTaskPool;
 use reth_trie::{
-    hashed_cursor::HashedPostStateCursorFactory, updates::StorageWriter, HashedPostState,
-    HashedStorage, StateRoot,
+    hashed_cursor::HashedPostStateCursorFactory, HashedPostState, HashedStorage, StateRoot,
 };
 use reth_trie_parallel::{async_root::AsyncStateRoot, parallel_root::ParallelStateRoot};
 use std::collections::HashMap;
@@ -31,7 +30,7 @@ pub fn calculate_state_root(c: &mut Criterion) {
             HashedStateChanges(db_state).write_to_db(provider_rw.tx_ref()).unwrap();
             let (_, updates) =
                 StateRoot::from_tx(provider_rw.tx_ref()).root_with_updates().unwrap();
-            updates.flush(&StorageWriter, provider_rw.tx_ref()).unwrap();
+            StorageWriter.write_trie_updates(updates, provider_rw.tx_ref()).unwrap();
             provider_rw.commit().unwrap();
         }
