+++ crates/trie/trie/src/trie.rs
@@ -548,7 +548,6 @@ mod tests {
     use crate::{
         prefix_set::PrefixSetMut,
         test_utils::{state_root, state_root_prehashed, storage_root, storage_root_prehashed},
-        updates::StorageWriter,
         BranchNodeCompact, TrieMask,
     };
     use proptest::{prelude::ProptestConfig, proptest};
@@ -559,7 +558,9 @@ mod tests {
         transaction::DbTxMut,
     };
     use reth_primitives::{hex_literal::hex, Account, StorageEntry, U256};
-    use reth_provider::{test_utils::create_test_provider_factory, DatabaseProviderRW};
+    use reth_provider::{
+        test_utils::create_test_provider_factory, DatabaseProviderRW, StorageWriter,
+    };
     use reth_trie_common::triehash::KeccakHasher;
     use std::{
         collections::{BTreeMap, HashMap},
@@ -622,7 +623,7 @@ mod tests {
         let modified_root = loader.root().unwrap();
 
         // Update the intermediate roots table so that we can run the incremental verification
-        trie_updates.flush(&StorageWriter, tx.tx_ref()).unwrap();
+        StorageWriter.write_trie_updates(trie_updates, tx.tx_ref()).unwrap();
 
         // 3. Calculate the incremental root
         let mut storage_changes = PrefixSetMut::default();
@@ -1224,7 +1225,7 @@ mod tests {
 
         let (got, updates) = StateRoot::from_tx(tx.tx_ref()).root_with_updates().unwrap();
         assert_eq!(expected, got);
-        updates.flush(&StorageWriter, tx.tx_ref()).unwrap();
+        StorageWriter.write_trie_updates(updates, tx.tx_ref()).unwrap();
 
         // read the account updates from the db
         let mut accounts_trie = tx.tx_ref().cursor_read::<tables::AccountsTrie>().unwrap();
@@ -1271,7 +1272,7 @@ mod tests {
                     state.iter().map(|(&key, &balance)| (key, (Account { balance, ..Default::default() }, std::iter::empty())))
                 );
                 assert_eq!(expected_root, state_root);
-                trie_updates.flush(&StorageWriter, tx.tx_ref()).unwrap();
+                StorageWriter.write_trie_updates(trie_updates, tx.tx_ref()).unwrap();
             }
         }
     }
